+++ cmd/telegraf/telegraf.go
@@ -313,7 +313,10 @@ func main() {
 			DisplayName: "Telegraf Data Collector Service",
 			Description: "Collects data using a series of plugins and publishes it to" +
 				"another series of plugins.",
-			Arguments: []string{"-config", "C:\\Program Files\\Telegraf\\telegraf.conf"},
+			Arguments: []string{
+			    "-config", "C:\\Program Files\\Telegraf\\telegraf.conf",
+			    "-config-directory", "C:\\Program Files\\Telegraf\\telegraf.d",
+			},
 		}
 
 		prg := &program{}