+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - [#2077](https://github.com/influxdata/telegraf/issues/2077): SQL Server Input - Arithmetic overflow error converting numeric to data type int.
 - [#2262](https://github.com/influxdata/telegraf/issues/2262): Flush jitter can inhibit metric collection.
+- [#1797](https://github.com/influxdata/telegraf/issues/1797): Windows service does not support config directory.
 
 ## v1.2 [2017-01-00]